Create sample_scanner
we use the sniff function from scapy to capture WiFi packets on the specified interface (wlan0mon in this case). We then define a callback function (scan_callback) that is called whenever a WiFi beacon or probe response packet is captured. This function extracts the network name, BSSID, and signal strength from the packet and prints them to the console.

Please note that the scapy library is quite powerful and flexible, but it can also be more complex to use than some of the other options. You may need to spend some time reading the documentation and examples to get a good understanding of how to use it effectively.